Move MEVIS adaptor to separate folder
